pyproject.toml
[tool.poetry]
name = "f1-bot"
version = "2.0.0"
description = "A Discord bot to view F1 stats."
authors = ["SmCTwelve <smctwelve@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "f1_bot"}]

[tool.poetry.dependencies]
python = "~3.11"
beautifulsoup4 = "^4.12.0"
lxml = "^4.9.2"
matplotlib = "^3.7.1"
python-dateutil = "^2.8.2"
tabulate = "^0.9.0"
py-cord = "^2.4.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
